+++ example/android/app/src/main/kotlin/com/example/smart_google_ads_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.smart_google_ads_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()